edge/requirements.txt
pydantic==2.10.4
requests==2.32.3
pyyaml==6.0.2
